refactor(amdsmi-tests): nest cpu/nic/ifoe C++ tests by feature
Give the smaller components the same <component>/<feature>/ layout as gpu/:

- ifoe/: move fabric_read.* -> ifoe/fabric/ and ifoe_info_read.* -> ifoe/identity/
  (update the two main.cc includes).
- cpu/: seed clock/ and power/ with placeholders (the two certain CPU features).
- nic/: seed discovery/ and identity/ with placeholders (its full taxonomy).

Document the convention in test-design.md: per-feature subdirectories, with a
suggested taxonomy mirroring the Python suite's test_<feature>.py files, and
refresh the directory tree. No CMake change needed thanks to CONFIGURE_DEPENDS.
