+++ projects/amdsmi/docs/conceptual/test-design.md
@@ -128,36 +128,36 @@ tests/amd_smi_test/
     │   ├── mutual_exclusion.{h,cc}
     │   └── sys_info_read.{h,cc}
     ├── cpu/
-    │   └── placeholder.cc           # Stub — CPU C++ tests added here
+    │   ├── clock/
+    │   │   └── placeholder.cc       # Stub — CPU clock C++ tests added here
+    │   └── power/
+    │       └── placeholder.cc       # Stub — CPU power C++ tests added here
     ├── nic/
-    │   └── placeholder.cc           # Stub — NIC and switch C++ tests added here
+    │   ├── discovery/
+    │   │   └── placeholder.cc       # Stub — NIC/switch discovery C++ tests added here
+    │   └── identity/
+    │       └── placeholder.cc       # Stub — NIC/switch identity C++ tests added here
     └── ifoe/
-        ├── fabric_read.{h,cc}       # IFoE fabric link reads
-        └── ifoe_info_read.{h,cc}    # IFoE endpoint info reads
+        ├── fabric/
+        │   └── fabric_read.{h,cc}   # IFoE fabric link reads
+        └── identity/
+            └── ifoe_info_read.{h,cc} # IFoE endpoint info reads
 ```
 
 ### Component subdirectory depth
 
-`gpu/` is split into per-feature subdirectories (`clock/`, `power/`, `metrics/`, …) because it
-carries many tests across many features. The smaller components start flatter and gain that depth
-only as they accumulate tests, so the tree never holds empty scaffolding:
+Each component groups tests into per-feature subdirectories (`<component>/<feature>/`, like `gpu/`).
+A feature gets its own subdirectory even for a single test. A `placeholder.cc` holds a feature
+directory until its first real test lands.
 
-- `cpu/` and `nic/` currently hold a single `placeholder.cc` — no C++ tests exist for them yet.
-- `ifoe/` keeps its two files flat for now.
+The names below are suggestions that mirror the Python suite's `test_<feature>.py` files. Sharing
+names lets a feature line up across both suites. Adapt them as the APIs warrant.
 
-When a component gains real tests, mirror the `gpu/` layout (`<component>/<feature>/`), using the
-**Python feature taxonomy as the canonical target** (the Python suite already splits these out, so
-the two suites stay aligned):
-
-| Component | Target feature subdirectories |
+| Component | Suggested feature subdirectories |
 | :--- | :--- |
 | `cpu/` | `clock/`, `dimm/`, `energy/`, `hsmp/`, `identity/`, `power/`, `thermal/` |
 | `nic/` | `discovery/`, `identity/` |
 | `ifoe/` | `fabric/`, `identity/` |
-
-Because `CMakeLists.txt` globs sources with `CONFIGURE_DEPENDS`, adding these subdirectories later
-needs no CMake change — drop the new `.cc` in the right place and rebuild. Promote a feature to its
-own subdirectory once it holds more than one file; a lone file can stay flat in the component root.
 
 ### Naming conventions
