+++ projects/amdsmi/tests/amd_smi_test/functional/ifoe/fabric/fabric_read.h
