+++ projects/amdsmi/tests/amd_smi_test/functional/ifoe/identity/ifoe_info_read.h
