+++ projects/amdsmi/tests/amd_smi_test/main.cc
@@ -58,8 +58,8 @@
 #include "functional/gpu/thermal/fan_read_write.h"
 #include "functional/gpu/thermal/temp_read.h"
 #include "functional/gpu/xgmi/xgmi_read_write.h"
-#include "functional/ifoe/fabric_read.h"
-#include "functional/ifoe/ifoe_info_read.h"
+#include "functional/ifoe/fabric/fabric_read.h"
+#include "functional/ifoe/identity/ifoe_info_read.h"
 #include "functional/system/cross_process_serialization.h"
 #include "functional/system/hw_topology_read.h"
 #include "functional/system/init_shutdown_refcount.h"
